Civilians loaded in ships will no longer be captured when a civilian is fought in a city
I finally came to the realization as to what was causing captured civilian units to appear on ships outside a city. It turns out the code didn't check for whether or not a civilian was loaded into a transport before marking them for capture. Now it checks if they're in a transport, and if it's a water transport, they will not be captured. Land transport-mounted civvies still get captured though.

+++ Project Files/DLLSources/CvUnit.cpp
@@ -1746,15 +1746,24 @@ void CvUnit::updateCombat(bool bQuick)
 				{
 					IDs.push_back(pUnitNode->m_data);
 
-					//Dyllin - Have to mark all civilians as captured before killing them due to strange capture bug that would move the units
-					//that were to be captured to a different tile despite also capturing them.
+					//Dyllin - Mark all civilians not loaded in ships as captured before killing them.
 					CvUnit* pLoopUnit = ::getUnit(pUnitNode->m_data);
 					if (pLoopUnit != NULL && pLoopUnit != pDefender)
 					{
 						if (isEnemy(pLoopUnit->getCombatTeam(getTeam(), pPlot), pPlot) &&
 							(pLoopUnit->isCapturableLandUnit() || pLoopUnit->isYield()))
 						{
-							pLoopUnit->setCapturingPlayer(getOwnerINLINE());
+							if (pLoopUnit->getTransportUnit() == NULL)
+							{
+								pLoopUnit->setCapturingPlayer(getOwnerINLINE());
+							}
+							else
+							{
+								if (pLoopUnit->getTransportUnit()->getDomainType() != DOMAIN_SEA)
+								{
+									pLoopUnit->setCapturingPlayer(getOwnerINLINE());
+								}
+							}
 						}
 					}
 
@@ -1773,7 +1782,17 @@ void CvUnit::updateCombat(bool bQuick)
 						if (isEnemy(pLoopUnit->getCombatTeam(getTeam(), pPlot), pPlot) &&
 							(pLoopUnit->isCapturableLandUnit() || pLoopUnit->isYield()))
 						{
-							pLoopUnit->kill(false);
+							if (pLoopUnit->getTransportUnit() == NULL)
+							{
+								pLoopUnit->kill(false);
+							}
+							else
+							{
+								if (pLoopUnit->getTransportUnit()->getDomainType() != DOMAIN_SEA)
+								{
+									pLoopUnit->kill(false);
+								}
+							}
 						}
 					}
 				}
